Fix missing grade for really bad performance results
When the results of the tests is worse than the expected the program
must assign an unknown grade to denotate that it does not knows how to
quality the responsiveness of the website. This usually never happens
but is a contra measure for when the testing servers or the website is
in maintenance mode.